Feat[#250] : MaskIndex
