+++ Server/kloudy/views.py
@@ -10,7 +10,7 @@ import datetime
 from pathlib import Path
 import environ
 from .networkAPI import *
-from .weathers import MainWeather, UmbrellaIndex
+from .weathers import MainWeather, UmbrellaIndex, MaskIndex
 
 CSV_PATH = './kloudy/csv/Locations.csv'
 
@@ -111,7 +111,7 @@ def time_interval_weather():
                 UmbrellaIndex.save_umbrella_hourly(umbrella_index, rains, location.code)
                 
             # 마스크 지수
-            mask_info = get_mask_index(air_jsonObject, flower_jsonObject)
+            mask_info = MaskIndex.get_mask_index(air_jsonObject, flower_jsonObject)
             if mask_info != [0, 0, 0, 0]:
                 if time % 2 != 0:
                     mask_index = MaskIndexEven.objects.filter(code = location.code).first()
@@ -239,7 +239,7 @@ def time_interval_weather():
             umbrella_index_even.save()
             UmbrellaIndex.save_umbrella_hourly(umbrella_index_odd, rains, location.code)
 
-            mask_info = get_mask_index(air_jsonObject, flower_jsonObject)
+            mask_info = MaskIndex.get_mask_index(air_jsonObject, flower_jsonObject)
             mask_status, pm25value, pm10value, pollen_index = mask_info
             print(f'마스크 지수: {mask_status}, {pm25value}, {pm10value}, {pollen_index}')
             mask_index_odd = MaskIndexOdd.objects.create(weather_index = weather_index_odd, code = location.code, status = mask_status, pm25value = pm25value, pm10value = pm10value, pollen_index = pollen_index)
@@ -293,50 +293,6 @@ def calculate_time():
     time = cal_time([now.strftime("%H"), now.strftime("%M")])
     
     return [today, time]
-
-def get_mask_index(air_jsonObject, flower_jsonObject):
-    if air_jsonObject.get('response').get('header').get('resultCode') != "00":
-        return [0, 0, 0, 0]
-    if flower_jsonObject.get('response').get('header').get('resultCode') != "00" and flower_jsonObject.get('response').get('header').get('resultCode') != "99" :
-        return [0, 0, 0, 0]
-
-    pm25_str = air_jsonObject.get('response').get('body').get('items')[0].get('pm25Value')
-    if pm25_str != "-":
-        pm25 = float(pm25_str)
-    else:
-        pm25 = 0
-
-    pm10_str = air_jsonObject.get('response').get('body').get('items')[0].get('pm10Value')
-    if pm10_str != "-":
-        pm10 = float(pm10_str)
-    else:
-        pm10 = 0
-
-    flower_quality = 0
-    if flower_jsonObject.get('response').get('header').get('resultCode') == '0':
-        flower_quality = int(flower_jsonObject.get('response').get('body').get('items').get('item').get('today'))
-
-    status = cal_mask_status(pm25)
-    pm25value = pm25
-    pm10value = pm10
-    pollen_index = flower_quality
-
-    return [status, pm25value, pm10value, pollen_index]
-
-# 0~15 : 좋음 0 / 16~35 : 보통 1 / 36~75 : 나쁨 2 / 76 : 매우나쁨 3
-def cal_mask_status(pm25):
-    result = 0
-
-    if 0 <= pm25 < 16:
-        result = 0
-    elif 16 <= pm25 < 36:
-        result = 1
-    elif 36 <= pm25 < 76:
-        result = 3
-    elif 76 <= pm25:
-        result = 4
-
-    return result
 
 # 일일 최저, 06시, 07시, 08시, 09시의 온도를 더해서 5로 나누어줌.
 def get_outer_index(weather_24h_jsonObject):
